pyproject.toml
[tool.poetry]
name = "memes"
version = "0.1.0"
description = ""
authors = ["Joao Vitor Maia <joaocampo2@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "^12.0"
memedict = "^1.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
